Added JWT generation to lookup handler
I’ve added the firebase JWT generation code to the lookup handler
instead of keeping it in the XAPI handler. The added code looks into
the context for the user’s id and then generates a JWT based upon the
uid. The JWT is returned as a part of a JSON object with the key “jwt.”